Still a lot to fix

+++ src/clients/rateapi/dto.go
@@ -1,7 +1,7 @@
 package rateapi
 
 type RateResponse struct {
-	OldAmount   int     `json:"old_amount"`
+	OldAmount   float64 `json:"old_amount"`
 	OldCurrency string  `json:"old_currency"`
 	NewCurrency string  `json:"new_currency"`
 	NewAmount   float64 `json:"new_amount"`
